+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/database/model/SessionRealmModule.kt
@@ -37,6 +37,7 @@ import io.realm.annotations.RealmModule
             SyncEntity::class,
             PendingThreePidEntity::class,
             UserEntity::class,
+            StringCountedEntity::class,
             IgnoredUserEntity::class,
             BreadcrumbsEntity::class,
             UserThreePidEntity::class,